Sidebar component
Tenker en sidebar  kan være smud, så legger den inn som en component og evt droppe Navbar
