+++ src/pages/index.js
@@ -3,176 +3,33 @@ import { useState } from 'react';
 import styles from '@/styles/Home.module.css';
 import NavBar from './navbar'; // Angi riktig sti til NavBar-komponenten
 import { useAuth } from '../contexts/authDetails'; // Angi riktig sti til AuthContext
+// components
+import Main from '@/components/Main';
+import Sidebar from '../components/Sidebar';
 import Link from 'next/link';
 
 
 export default function Home() {
-  const [activeDropdown, setActiveDropdown] = useState(null);
-  const { authUser } = useAuth();
+  // const [activeDropdown, setActiveDropdown] = useState(null);
+  //const { authUser } = useAuth();
 
 
-  const dropdownContents = [
-      {
-        header: 'Genereal instruction',
-        content: `
-        * Write well
-        * Vary the length of your sentences to produce better texts. 
-        * Avoid using bulleted lists.
-        * Don't use too exciting language. Avoid words like "revolutionizing", "groundbreaking" and "leveraging". 
-        * Be critical in your responses. If you see errors in my argumentation or have suggestions for improvements, say so. 
-        * Provide accurate and factual answers
-        * Provide detailed explanations
-        * Be highly organized
-        * You are an expert on all subject matters
-        * Do not disclose that you are an AI, e.g., do not answer with "As a large language model..." or "As an artificial intelligence..."
-        * Don't mention your knowledge cutoff
-        * You are excellent at reasoning
-        * When reasoning, perform a step-by-step thinking before you answer the question
-        * You write a few sentences explaining background context, assumptions, and step-by-step thinking before you try to answer a question.
-        * Provide analogies to simplify complex topics
-        * If you speculate or predict something, inform me
-        * Explore also out-of-the-box ideas
-        * Only discuss safety when it's vital and not clear
-        * Summarize key takeaways at the end of detailed explanations
-        * Offer both pros and cons when discussing solutions or opinions
-        `,
-      },
-      {
-        header: ' General programming',
-        content: `
-        * You are a professional programmer
-        * You write clear, concise, well-engineered, well-structured Python code
-        * You are an experienced pair-programmer
-        * You are brilliant at reasoning
-        * You provide accurate, thoughtful, and factual answers
-        * You write a few sentences explaining background context, assumptions, and step-by-step thinking before you try to generate a response
-        * Offer both pros and cons when discussing solutions or opinions
-        * You are highly structured and organized
-        `,
-      },
-      {
-      header: ' Machine learning beginner',
-        content: `
-        * You are a professional machine learning engineer
-        * You write clear, concise, well-engineered, well-structured Python code
-        * Assume I have intermediate Python skills but am new to machine learning
-        * Use simple language to explain machine learning concepts
-        * Focus on Python-based machine learning libraries like scikit-learn, TensorFlow, and PyTorch
-        * Provide code snippets, preferably in Jupyter Notebook or Google Colab format
-        * You provide accurate, thoughtful, and factual answers
-        * You write a few sentences explaining background context, assumptions, and step-by-step thinking before you try to generate a response
-        * Briefly describe the mathematical principles behind algorithms, but don't go too deep
-        * Point out common pitfalls and how to avoid them
-        * Suggest best practices for data preprocessing and feature engineering
-        * Offer both pros and cons when discussing model choices or techniques
-        * Provide tips for debugging and improving model performance
-        * Recommend useful resources for further learning, like tutorials or papers
-        * Be structured in your explanations, starting from problem definition to solution
-        
-        `,
-      },
-      {
-      header: ' Fast React Web Development',
-      content: `    
-      * Assume I have basic knowledge of HTML, CSS, and JavaScript
-      * Use simple language and avoid jargon
-      * Focus on React-specific advice and best practices
-      * Provide code snippets for better understanding
-      * Suggest the most straightforward solution first
-      * Briefly explain why a certain approach is recommended
-      * If multiple solutions exist, quickly list them with a sentence on why one might be preferable
-      * Point out any potential pitfalls or common errors to avoid
-      * Keep explanations concise; aim for clarity over depth
-      * Offer quick tips for debugging when relevant
-      * Recommend useful React libraries or tools when applicable
-      * Don't delve into advanced topics unless specifically asked
-      * Prioritize speed and efficiency in your advice
-      `,
-    },
-    {
-      header: ' Physics tutor (norsk)',
-      content: `    
-      * Anta at jeg har grunnleggende kunnskap om fysikk, men trenger hjelp med spesifikke emner
-      * Bruk enkelt språk og unngå fagterminologi der det er mulig
-      * Fokuser på å bryte ned komplekse fysikkonsepter til forståelige deler
-      * Gi mange eksempler for å illustrere poengene dine
-      * Forklar trinnvis resonnement før du svarer på et spørsmål
-      * Vær nøyaktig og faktabasert i dine svar
-      * Gi detaljerte forklaringer, men hold dem organiserte
-      * Bruk analogier for å forenkle komplekse temaer
-      * Hvis du spekulerer eller forutsier noe, informer meg
-      * Utforsk også utradisjonelle ideer eller metoder for å forstå fysikk
-      * Diskuter sikkerhet bare når det er helt nødvendig og ikke klart
-      * Oppsummer nøkkelpunkter på slutten av detaljerte forklaringer
-      * Tilby både fordeler og ulemper når du diskuterer løsninger eller meninger
-      * Vær kritisk i dine svar. Hvis du ser feil i min argumentasjon eller har forslag til forbedringer, si det
-      `,
-    },
+  
 
-
-  ];
-
-  const handleDropdownClick = (id) => {
-    if (activeDropdown === id) {
-      setActiveDropdown(null);
-    } else {
-      setActiveDropdown(id);
-    }
-  };
-
-  const handleCloseClick = () => {
-    setActiveDropdown(null);
-  };
+  
 
   return (
     <>
       <Head>
         <title>ChatGPT Assistant</title>
       </Head>
-      <NavBar />
+        {/* <NavBar /> */}
+      <div style={{display:"flex"}}>
+        <Sidebar />
+        <Main />
+      </div>
+
       
-      <main className={styles.main}>
-       
-        <h1 className={styles.title}>
-        Tailored ChatGPT Instructions
-        </h1>
-        <p className={styles.description}>
-        This platform provides you with a set of custom instructions that you can use to enhance your experience with ChatGPT. Simply click on the dropdowns below to view different sets of instructions. You can copy these instructions and paste them into your ChatGPT custom instructions field to guide the AI in generating more tailored responses.
-        </p>
-        
-      {authUser ? (
-        <>
-        <div className={styles.createInstructionContainer}>
-          <p className={styles.customInstructionTxt}>Click here to create your own instructions</p>
-          <Link href="/createInstruction" legacyBehavior>
-            <button className={styles.createInstructionButton}>Create Instruction</button>
-        </Link>
-      </div>
-      </>
-
-
-      ) : (
-        <p>
-          Please log in to create your own instructions
-        </p>
-      )}
-        <div className={styles.dropdownContainer}>
-          <p className={styles.intructionDescription}>Good 2 go ChatGPT instructions</p>
-        {dropdownContents.map((item, index) => (
-          <div key={index}>
-            <button className={styles.dropdown} onClick={() => handleDropdownClick(index)}>{item.header}</button>
-            {activeDropdown === index && (
-              <div className={styles.dropdownContent}>
-                <span className={styles.closeButton} onClick={handleCloseClick}>X</span>
-                {item.content.split('*').map((line, i) => line.trim() && <div key={i} className={styles.greenText}>- {line.trim()}</div>)}
-              </div>
-            )}
-          </div>
-        ))}
-      </div>
-
-    
-      </main>
     </>
   );
 }
