+++ src/styles/Home.module.css
@@ -1,6 +1,7 @@
 
 /* Main container */
 .main {
+  flex: 3;
   display: flex;
   flex-direction: column;
   align-items: center;
