+++ src/styles/globals.css
@@ -1,4 +1,8 @@
+
+
 body {
+  margin: 0;
+  padding: 0;
     background-color: black;
     color: white;
     font-family: 'Arial', sans-serif; /* Du kan endre denne til en mer moderne eller dystopisk font om du ønsker */
